requirements.txt
streamlit
requests
beautifulsoup4
pandas
plotly
google.generativeai
python-dotenv
